Add option to build for /usr/local
Very help if you build fcitx5 and install to /usr/local

+++ fcp-zmq/ims/dev.sh
@@ -8,6 +8,14 @@ if [ $1 == "init_cmake" ]; then
     exit 0
 fi
 
+if [ $1 == "init_cmake_local" ]; then
+    mkdir -p build
+    cd ./build
+    cmake -DCMAKE_EXPORT_COMPILE_COMMANDS=1 -DCMAKE_BUILD_TYPE=Debug ..
+    cd ../
+    exit 0
+fi
+
 if [ $1 == "clean" ]; then
     if [ -d ./build ]; then
         rm -rf ./build
